NLP
Read up on any innovative technology using NLP (by companies such as
Google or IBM, for instance) and write a brief summary about the
technology, what it achieves/does, and an overview of how it works (250 -
500 words).
To take an example, you may have noticed Gmail’s auto-response
suggestions on your incoming emails. If I send an email to your Gmail
address asking for an appointment, on opening the mail you would notice
Gmail’s automatically suggested response options such as “Yes, that works
for me” and “Sorry, I’m not available at that time. Save your answer in a file.